+++ tests/Point_Test.cpp
@@ -6,69 +6,69 @@
 TEST(PointTests, constructor_default) {
 	Point p1;
 	EXPECT_EQ(typeid(p1).name(), typeid(Point).name());
-	EXPECT_FLOAT_EQ(p1.getX(), 0.0f);
-	EXPECT_FLOAT_EQ(p1.getY(), 0.0f);
-	EXPECT_FLOAT_EQ(p1.getZ(), 0.0f);
-	EXPECT_FLOAT_EQ(p1.getW(), 1.0f);
+	EXPECT_FLOAT_EQ(p1.getX(), 0.0);
+	EXPECT_FLOAT_EQ(p1.getY(), 0.0);
+	EXPECT_FLOAT_EQ(p1.getZ(), 0.0);
+	EXPECT_FLOAT_EQ(p1.getW(), 1.0);
 }
 
 TEST(PointTests, constructor_surcharged) {
-	Point p1(1.0f);
+	Point p1(1.0);
 	EXPECT_EQ(typeid(p1).name(), typeid(Point).name());
-	EXPECT_FLOAT_EQ(p1.getX(), 1.0f);
-	EXPECT_FLOAT_EQ(p1.getY(), 0.0f);
-	EXPECT_FLOAT_EQ(p1.getZ(), 0.0f);
-	EXPECT_FLOAT_EQ(p1.getW(), 1.0f);
+	EXPECT_FLOAT_EQ(p1.getX(), 1.0);
+	EXPECT_FLOAT_EQ(p1.getY(), 0.0);
+	EXPECT_FLOAT_EQ(p1.getZ(), 0.0);
+	EXPECT_FLOAT_EQ(p1.getW(), 1.0);
 
-	Point p2(1.0f, 2.0f);
+	Point p2(1.0, 2.0);
 	EXPECT_EQ(typeid(p2).name(), typeid(Point).name());
-	EXPECT_FLOAT_EQ(p2.getX(), 1.0f);
-	EXPECT_FLOAT_EQ(p2.getY(), 2.0f);
-	EXPECT_FLOAT_EQ(p2.getZ(), 0.0f);
-	EXPECT_FLOAT_EQ(p2.getW(), 1.0f);
+	EXPECT_FLOAT_EQ(p2.getX(), 1.0);
+	EXPECT_FLOAT_EQ(p2.getY(), 2.0);
+	EXPECT_FLOAT_EQ(p2.getZ(), 0.0);
+	EXPECT_FLOAT_EQ(p2.getW(), 1.0);
 
-	Point p3(1.0f, 2.0f, 3.0f);
+	Point p3(1.0, 2.0, 3.0);
 	EXPECT_EQ(typeid(p3).name(), typeid(Point).name());
-	EXPECT_FLOAT_EQ(p3.getX(), 1.0f);
-	EXPECT_FLOAT_EQ(p3.getY(), 2.0f);
-	EXPECT_FLOAT_EQ(p3.getZ(), 3.0f);
-	EXPECT_FLOAT_EQ(p3.getW(), 1.0f);
+	EXPECT_FLOAT_EQ(p3.getX(), 1.0);
+	EXPECT_FLOAT_EQ(p3.getY(), 2.0);
+	EXPECT_FLOAT_EQ(p3.getZ(), 3.0);
+	EXPECT_FLOAT_EQ(p3.getW(), 1.0);
 }
 
 TEST(PointTests, constructor_copy) {
-	Point p1(1.0f, 2.0f, 3.0f);
+	Point p1(1.0, 2.0, 3.0);
 	Point p2(p1);
 	EXPECT_EQ(typeid(p2).name(), typeid(Point).name());
-	EXPECT_FLOAT_EQ(p2.getX(), 1.0f);
-	EXPECT_FLOAT_EQ(p2.getY(), 2.0f);
-	EXPECT_FLOAT_EQ(p2.getZ(), 3.0f);
-	EXPECT_FLOAT_EQ(p2.getW(), 1.0f);
+	EXPECT_FLOAT_EQ(p2.getX(), 1.0);
+	EXPECT_FLOAT_EQ(p2.getY(), 2.0);
+	EXPECT_FLOAT_EQ(p2.getZ(), 3.0);
+	EXPECT_FLOAT_EQ(p2.getW(), 1.0);
 }
 
 TEST(PointTests, constructor_moving) {
 	Point p1;
-	p1 = Point(1.0f, 2.0f, 3.0f);
+	p1 = Point(1.0, 2.0, 3.0);
 	EXPECT_EQ(typeid(p1).name(), typeid(Point).name());
-	EXPECT_FLOAT_EQ(p1.getX(), 1.0f);
-	EXPECT_FLOAT_EQ(p1.getY(), 2.0f);
-	EXPECT_FLOAT_EQ(p1.getZ(), 3.0f);
-	EXPECT_FLOAT_EQ(p1.getW(), 1.0f);
+	EXPECT_FLOAT_EQ(p1.getX(), 1.0);
+	EXPECT_FLOAT_EQ(p1.getY(), 2.0);
+	EXPECT_FLOAT_EQ(p1.getZ(), 3.0);
+	EXPECT_FLOAT_EQ(p1.getW(), 1.0);
 }
 
 TEST(PointTests, assignment) {
-	Point p1(1.0f, 2.0f, 3.0f);
+	Point p1(1.0, 2.0, 3.0);
 	Point p2 = p1;
 	EXPECT_EQ(typeid(p2).name(), typeid(Point).name());
-	EXPECT_FLOAT_EQ(p2.getX(), 1.0f);
-	EXPECT_FLOAT_EQ(p2.getY(), 2.0f);
-	EXPECT_FLOAT_EQ(p2.getZ(), 3.0f);
-	EXPECT_FLOAT_EQ(p2.getW(), 1.0f);
+	EXPECT_FLOAT_EQ(p2.getX(), 1.0);
+	EXPECT_FLOAT_EQ(p2.getY(), 2.0);
+	EXPECT_FLOAT_EQ(p2.getZ(), 3.0);
+	EXPECT_FLOAT_EQ(p2.getW(), 1.0);
 }
 
 TEST(PointTests, equality) {
-	Point p1(1.0f, 2.0f, 3.0f);
-	Point p2(1.0f, 2.0f, 3.0f);
-	Point p3(3.0f, 2.0f, 1.0f);
+	Point p1(1.0, 2.0, 3.0);
+	Point p2(1.0, 2.0, 3.0);
+	Point p3(3.0, 2.0, 1.0);
 
 	EXPECT_TRUE(p1 == p1);
 	EXPECT_FALSE(p1 != p1);
@@ -135,10 +135,10 @@ TEST(PointTests, negate_point) {
 }
 
 TEST(PointTests, scalar_multiplication) {
-	float s1 = 2.0f;
-	float s2 = 0.25f;
-	Point p1(1.0f, 2.0f, 3.0f);
-	Point ref1(2.0f, 4.0f, 6.0f);
+	double s1 = 2.0;
+	double s2 = 0.25;
+	Point p1(1.0, 2.0, 3.0);
+	Point ref1(2.0, 4.0, 6.0);
 	Point p2 = p1 * s1;
 	EXPECT_EQ(typeid(p2).name(), typeid(Point).name());
 
@@ -146,16 +146,16 @@ TEST(PointTests, scalar_multiplication) {
 	EXPECT_FLOAT_EQ(p2.getX(), ref1.getX());
 	EXPECT_FLOAT_EQ(p2.getY(), ref1.getY());
 	EXPECT_FLOAT_EQ(p2.getZ(), ref1.getZ());
-	EXPECT_FLOAT_EQ(p2.getW(), 2.0f);
+	EXPECT_FLOAT_EQ(p2.getW(), 2.0);
 
 	Point p3 = p2 * s2;
 	EXPECT_EQ(typeid(p3).name(), typeid(Point).name());
-	Point ref2(0.5f, 1.0f, 1.5f);
+	Point ref2(0.5, 1.0, 1.5);
 	EXPECT_TRUE(p3 != ref2);
 	EXPECT_FLOAT_EQ(p3.getX(), ref2.getX());
 	EXPECT_FLOAT_EQ(p3.getY(), ref2.getY());
 	EXPECT_FLOAT_EQ(p3.getZ(), ref2.getZ());
-	EXPECT_FLOAT_EQ(p3.getW(), 0.5f);
+	EXPECT_FLOAT_EQ(p3.getW(), 0.5);
 
 	p1 *= s1;
 	EXPECT_EQ(typeid(p1).name(), typeid(Point).name());
@@ -163,7 +163,7 @@ TEST(PointTests, scalar_multiplication) {
 	EXPECT_FLOAT_EQ(p1.getX(), ref1.getX());
 	EXPECT_FLOAT_EQ(p1.getY(), ref1.getY());
 	EXPECT_FLOAT_EQ(p1.getZ(), ref1.getZ());
-	EXPECT_FLOAT_EQ(p1.getW(), 2.0f);
+	EXPECT_FLOAT_EQ(p1.getW(), 2.0);
 
 	p2 *= s2;
 	EXPECT_EQ(typeid(p2).name(), typeid(Point).name());
@@ -171,7 +171,7 @@ TEST(PointTests, scalar_multiplication) {
 	EXPECT_FLOAT_EQ(p2.getX(), ref2.getX());
 	EXPECT_FLOAT_EQ(p2.getY(), ref2.getY());
 	EXPECT_FLOAT_EQ(p2.getZ(), ref2.getZ());
-	EXPECT_FLOAT_EQ(p2.getW(), 0.5f);
+	EXPECT_FLOAT_EQ(p2.getW(), 0.5);
 }
 
 TEST(PointTests, matrix_multiplication) {
@@ -179,8 +179,8 @@ TEST(PointTests, matrix_multiplication) {
 			2, 4, 4, 2,
 			8, 6, 4, 1,
 			0, 0, 0, 1);
-	Point p1(1.0f, 2.0f, 3.0f);
-	Point ref1(18.0f, 24.0f, 33.0f);
+	Point p1(1.0, 2.0, 3.0);
+	Point ref1(18.0, 24.0, 33.0);
 	Point p2 = m1 * p1;
 	EXPECT_EQ(typeid(p2).name(), typeid(Point).name());
 	EXPECT_TRUE(p2 == ref1);
